Make standard zip filesystems uninterruptible
